+++ packages/share_plus/share_plus/lib/share_plus.dart
@@ -7,7 +7,7 @@ import 'dart:ui';
 
 import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
 
-export 'package:share_plus_platform_interface/share_plus_platform_interface.dart' show ShareResult, ShareResultStatus, ShareWithAppWindows;
+export 'package:share_plus_platform_interface/share_plus_platform_interface.dart' show ShareResult, ShareResultStatus, ShareWithApp;
 
 /// Plugin for summoning a platform share sheet.
 class Share {
@@ -62,6 +62,10 @@ class Share {
   /// origin rect for the share sheet to popover from on iPads. It has no effect
   /// on non-iPads.
   ///
+  /// The optional `appName` parameter can be used to specify a default app or specific
+  /// app name from [ShareWithApp] for the share. Currently [appName] is supported for Windows platform app
+  /// only.
+  ///
   /// May throw [PlatformException] or [FormatException]
   /// from [MethodChannel].
   static Future<void> shareFiles(
@@ -70,7 +74,8 @@ class Share {
     String? subject,
     String? text,
     Rect? sharePositionOrigin,
-  }) {
+    ShareWithApp? appName
+      }) {
     assert(paths.isNotEmpty);
     assert(paths.every((element) => element.isNotEmpty));
     return _platform.shareFiles(
@@ -79,6 +84,7 @@ class Share {
       subject: subject,
       text: text,
       sharePositionOrigin: sharePositionOrigin,
+      appName: appName,
     );
   }