+++ packages/share_plus/share_plus_windows/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 3.0.1
 
 - Add, Share file with Windows desktop app
-- Add, enum `ShareWithAppWindows` to for selective app sharing in windows desktop
+- Add, enum `ShareWithApp` to for selective app sharing in windows desktop
 
 ## 3.0.0